% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{anova_linear_regression}
\alias{anova_linear_regression}
\title{다중선형회귀모형 분산분석표.}
\usage{
anova_linear_regression(.fit)
}
\arguments{
\item{.fit}{회귀모형 추정 결과.}
}
\value{
분산분석표 데이터프레임.
}
\description{
추정된 회귀모형을 이용하여 회귀성 검정을 위한 분산분석표를 생성한다.
}
\examples{
data(biometric, package = "dmtr")
fit <- fit_linear_regression(biometric, weight, c(age, height))
anova_linear_regression(fit)

}
