% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{austenbooks}
\alias{austenbooks}
\title{Jane Austen Books.}
\format{
2개의 열을 지닌 데이터 프레임:
\describe{
  \item{text}{텍스트}
  \item{book}{소설 제목}
}
}
\source{
\code{\link[janeaustenr]{austen_books}}
}
\usage{
austenbooks
}
\description{
Jane Austen의 6개의 소설의 텍스트를 저장한 데이터 프레임.
각 소설 텍스트는 약 70개의 문자 단위로 나뉘어 행을 구성한다.
}
\keyword{datasets}
