% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryclass1_test}
\alias{binaryclass1_test}
\title{두 범주 테스트 표본 1.}
\format{
2개의 행과 3개의 열을 지닌 데이터 프레임:
\describe{
  \item{id}{객체 고유번호}
  \item{x1, x2}{두 변수}
}
}
\usage{
binaryclass1_test
}
\description{
두 범주 학습표본 1 (\code{\link{binaryclass1_train}})을 이용해 학습된 분류규칙을 적용하기 위한 새 표본
}
\keyword{datasets}
