% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryclass2}
\alias{binaryclass2}
\title{두 범주 학습표본 2.}
\format{
9개의 행과 4개의 열을 지닌 데이터 프레임:
\describe{
  \item{id}{객체 고유번호}
  \item{x1, x2}{두 변수}
  \item{class}{범주 (1 혹은 2)}
}
}
\usage{
binaryclass2
}
\description{
두 개의 범주값과 두 개의 변수를 지닌 학습표본.
}
\keyword{datasets}
