% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryclass3_test}
\alias{binaryclass3_test}
\title{두 범주 테스트 표본 3.}
\format{
6개의 행과 3개의 열을 지닌 데이터 프레임:
\describe{
  \item{x1, x2}{두 변수}
  \item{class}{범주 (1 혹은 2)}
}
}
\usage{
binaryclass3_test
}
\description{
두 범주 학습표본 3 (\code{\link{binaryclass3_train}})을 이용해 학습된 분류규칙을 적용하기 위한 새 표본
}
\keyword{datasets}
