% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryclass3_train}
\alias{binaryclass3_train}
\title{두 범주 학습표본 표본 3.}
\format{
10개의 행과 3개의 열을 지닌 데이터 프레임:
\describe{
  \item{x1, x2}{두 변수}
  \item{class}{범주 (1 혹은 2)}
}
}
\usage{
binaryclass3_train
}
\description{
두 개의 범주값과 두 개의 변수를 지닌 학습표본.
}
\keyword{datasets}
