% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryclass4_inseparable}
\alias{binaryclass4_inseparable}
\title{두 범주 학습표본 표본 4 - 선형분리불가능.}
\format{
10개의 행과 3개의 열을 지닌 데이터 프레임:
\describe{
  \item{x1, x2}{두 변수}
  \item{class}{범주 (1 혹은 -1)}
}
}
\usage{
binaryclass4_inseparable
}
\description{
두 개의 범주값과 두 개의 변수를 지닌 학습표본. 해당 표본은 선형 분류기법으로는 두 범주가 완전히 분리되지 않는다.
}
\keyword{datasets}
