% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryclass4_separable}
\alias{binaryclass4_separable}
\title{두 범주 학습표본 표본 4 - 선형분리가능.}
\format{
9개의 행과 3개의 열을 지닌 데이터 프레임:
\describe{
  \item{x1, x2}{두 변수}
  \item{class}{범주 (1 혹은 -1)}
}
}
\usage{
binaryclass4_separable
}
\description{
두 개의 범주값과 두 개의 변수를 지닌 학습표본. 해당 표본은 선형 분류기법으로 두 범주가 완전히 분리된다.
}
\keyword{datasets}
