% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryclass5}
\alias{binaryclass5}
\title{두 범주 학습표본 표본 5.}
\format{
10개의 행과 2개의 열을 지닌 데이터 프레임:
\describe{
  \item{x}{변수}
  \item{y}{범주 (0 혹은 1)}
}
}
\usage{
binaryclass5
}
\description{
두 개의 범주값과 하나의 변수를 지닌 학습표본.
}
\keyword{datasets}
