% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biometric}
\alias{biometric}
\title{나이, 키, 몸무게 학습표본.}
\format{
10개의 행과 3개의 열을 지닌 데이터 프레임:
\describe{
  \item{age}{나이}
  \item{height}{키(cm)}
  \item{weight}{몸무게(kg)}
}
}
\usage{
biometric
}
\description{
나이와 키로써 몸무게를 설명하는 회귀모형을 추정하기 위한 학습표본.
}
\keyword{datasets}
