% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusterdata4}
\alias{clusterdata4}
\title{군집분석 데이터 4.}
\format{
8개의 행과 3개의 열을 지닌 데이터 프레임:
\describe{
  \item{id}{객체 고유번호}
  \item{x1, x2}{두 변수}
}
}
\usage{
clusterdata4
}
\description{
두 개의 실수형 변수로 이루어진 객체 데이터
}
\keyword{datasets}
