% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customerclass_test}
\alias{customerclass_test}
\title{고객 새 표본.}
\format{
1개의 행과 2개의 열을 지닌 데이터 프레임:
\describe{
  \item{x1}{성별 (남 혹은 여)}
  \item{x2}{나이 (10대, 20대, 30대, 40대)}
}
}
\usage{
customerclass_test
}
\description{
고객 학습표본 \code{\link{customerclass_train}}을 이용해 학습된 분류규칙을 적용하여 분류를 예측하고자 하는 새 고객 표본.
}
\keyword{datasets}
