% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customerclass_train}
\alias{customerclass_train}
\title{고객 학습표본.}
\format{
9개의 행과 4개의 열을 지닌 데이터 프레임:
\describe{
  \item{id}{고객 고유번호}
  \item{x1}{성별 (남 혹은 여)}
  \item{x2}{나이 (10대, 20대, 30대, 40대)}
  \item{y}{범주 (1 혹은 2)}
}
}
\usage{
customerclass_train
}
\description{
두 개의 범주값과 두 개의 변수를 지닌 학습표본.
}
\keyword{datasets}
