% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defecttype}
\alias{defecttype}
\title{공정변수-불량 종류 데이터}
\format{
18개의 행과 4개의 열을 지닌 데이터 프레임:
\describe{
  \item{id}{관측 고유번호}
  \item{x1, x2}{두 공정변수}
  \item{y}{불량 종류 (1, 2, 3)}
}
}
\usage{
defecttype
}
\description{
3종류의 불량을 2개의 공정변수로 예측하는 모형을 학습하기 위한 데이터.
}
\keyword{datasets}
