% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-metrics.R
\name{eval_mad}
\alias{eval_mad}
\title{평균절대오차}
\usage{
eval_mad(.y, .y_hat)
}
\arguments{
\item{.y}{관측 종속변수 값. 숫자형 벡터.}

\item{.y_hat}{예측 종속변수 값. 숫자형 벡터.}
}
\value{
평균절대오차.
}
\description{
예측값의 평균절대오차를 계산한다.
}
