% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-metrics.R
\name{eval_rmse}
\alias{eval_rmse}
\title{평균제곱오차의 제곱근}
\usage{
eval_rmse(.y, .y_hat)
}
\arguments{
\item{.y}{관측 종속변수 값. 숫자형 벡터.}

\item{.y_hat}{예측 종속변수 값. 숫자형 벡터.}
}
\value{
평균제곱오차의 제곱근.
}
\description{
예측값의 평균제곱오차의 제곱근을 계산한다.
}
