% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{evaluate_linear_regression}
\alias{evaluate_linear_regression}
\title{다중회귀모형 선택척도 산출}
\usage{
evaluate_linear_regression(.data, .yvar, .xvar)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.yvar}{종속변수.}

\item{.xvar}{완전모형에 속할 독립변수. 독립변수가 여러 개일 때는 벡터 형태로 제공한다. (e.g. \code{c(age, height)})}
}
\value{
\code{.xvar}의 각 변수조합에 대한 선택척도 데이터프레임.
}
\description{
회귀모형의 각 변수조합에 대한 선택척도를 산출한다.
}
\examples{
data(biometric, package = "dmtr")
evaluate_linear_regression(biometric, weight, c(age, height))

}
\keyword{regression}
