% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{financials}
\alias{financials}
\title{국내 증권회사의 주요 재무제표}
\format{
18개의 행과 6개의 열을 지닌 데이터 프레임:
\describe{
  \item{company}{증권회사명}
  \item{roa}{총자본 순이익율}
  \item{roe}{자기자본 순이익율}
  \item{bis}{자기자본비율}
  \item{de_ratio}{부채비율}
  \item{turnover}{자기자본 회전율}
}
}
\usage{
financials
}
\description{
국내 18개 증권회사의 주요 재무제표
}
\keyword{datasets}
