% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant-functions.R
\name{fisher_ld_prediction}
\alias{fisher_ld_prediction}
\title{피셔 선형 판별 함수 예측값.}
\usage{
fisher_ld_prediction(.w, .z, .newdata, ..., .levels = c(1L, 2L))
}
\arguments{
\item{.w}{선형 함수 계수.}

\item{.z}{선형 함수 분류 경계치.}

\item{.newdata}{새 데이터.}

\item{...}{범주 분류에 사용될 변수.}

\item{.levels}{범주.}
}
\value{
분류 예측값을 포함한 데이터 프레임.
}
\description{
새로운 데이터에 대해 피셔 선형 판별함수를 이용하여 범주를 분류한다.
}
\examples{
data(binaryclass2, package = "dmtr")
w <- fisher_ld(binaryclass2, class, x1, x2)
z <- fisher_ld_threshold(binaryclass2, class, x1, x2)
pred <- fisher_ld_prediction(w, z, binaryclass2, x1, x2, .levels = attr(w, "group"))

}
