% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{fit_binary_logistic_regression}
\alias{fit_binary_logistic_regression}
\title{이분 로지스틱 회귀분석 계수 추정.}
\usage{
fit_binary_logistic_regression(
  .data,
  .group_var,
  .xvar,
  .reflevel = NULL,
  .control = list(fnscale = -1, maxit = 10000)
)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.group_var}{범주변수.}

\item{.xvar}{범주 분류에 사용될 변수.}

\item{.reflevel}{이분 범주에서 0(negative) 값을 지닐 범주. NULL값일 때는 \code{.group_var} 범주
레벨에서 마지막 레벨을 사용한다.}

\item{.control}{\code{\link[stats]{optim}} 함수 수행 시 \code{control} 파라미터 리스트.}
}
\value{
리스트. 최우추정 계수 벡터 \code{betas}. 헤시안 행렬 \code{hessian}.
}
\description{
주어진 데이터에 대해 이분 로지스틱 회귀모형 계수를 추정한다.
}
\examples{
data(student, package = "dmtr")
fit_binary_logistic_regression(student, y, c(x1, x2, x3), .reflevel = "우수")

}
\keyword{logistic-regression}
