% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{다중회귀모형 선형 회귀계수 추정}
\usage{
fit_linear_regression(.data, .yvar, .xvar)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.yvar}{종속변수.}

\item{.xvar}{독립변수. 독립변수가 여러 개일 때는 벡터 형태로 제공한다. (e.g. \code{c(age, height)})}
}
\value{
리스트. 최우추정 계수 벡터 \code{betas}. 헤시안 행렬 \code{hessian}.
}
\description{
주어진 데이터에 대하여 다중회귀모형을 추정한다.
}
\examples{
data(biometric, package = "dmtr")
fit <- fit_linear_regression(biometric, weight, c(age, height))

}
