% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal-component-analysis.R
\name{fit_pca}
\alias{fit_pca}
\title{주성분 분해}
\usage{
fit_pca(.data, .xvar = everything(), .pc = NULL, .center = TRUE, .scale = TRUE)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.xvar}{독립변수. 독립변수가 여러 개일 때는 벡터 형태로 제공한다. (e.g. \code{c(age, height)}).
Default: \code{everthing()}, \code{.data}의 모든 변수에 대한 주성분 분해를 수행한다.}

\item{.pc}{추출할 주성분의 수.}

\item{.center}{평균조정 여부. TRUE이면 각 독립변수에 대해 평균조정을 수행한다.}

\item{.scale}{분산조정 여부. TRUE이면 각 독립변수에 대해 분산조정을 수행한다.}
}
\value{
리스트. 고유치 \code{eig}, 주성분행렬 \code{score},
        상관로딩 \code{loading}, 주성분 수 \code{ncomp}, 주성분 변동 기여율 \code{R2},
        평균조정 시 평균벡터 \code{center}, 분산조정 시 표본표준편차 벡터 \code{scale}.
}
\description{
주어진 데이터에 대하여 주성분을 분해한다.
}
\examples{
data(biometric, package = "dmtr")
fit1 <- fit_pca(biometric, c(age, height, weight), .pc = 2L)
fit2 <- fit_pca(biometric, .pc = 2L)

}
\keyword{principal-component-analysis}
