% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal-component-analysis.R
\name{fit_pcr}
\alias{fit_pcr}
\title{주성분 회귀분석}
\usage{
fit_pcr(.data, .yvar, .xvar, .pc = NULL, .center = TRUE, .scale = TRUE)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.yvar}{종속변수.}

\item{.xvar}{독립변수. 독립변수가 여러 개일 때는 벡터 형태로 제공한다. (e.g. \code{c(age, height)}).}

\item{.pc}{추출할 주성분의 수.}

\item{.center}{평균조정 여부. TRUE이면 각 독립변수에 대해 평균조정을 수행한다.}

\item{.scale}{분산조정 여부. TRUE이면 각 독립변수에 대해 분산조정을 수행한다.}
}
\value{
리스트. 주성분을 독립변수로 하여 수행한 \code{fit_linear_regression()}의 결과값.
}
\description{
주어진 데이터에 대하여 주성분 회귀모형을 추정한다.
}
\examples{
data(biometric, package = "dmtr")
fit <- fit_pcr(biometric, weight, c(age, height), .pc = 1L)

}
\keyword{principal-component-analysis}
