% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-least-squares.R
\name{fit_plsr}
\alias{fit_plsr}
\title{부분최소자승 회귀분석}
\usage{
fit_plsr(.data, .yvar, .xvar, .ncomp = NULL, .center = TRUE, .scale = TRUE)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.yvar}{종속변수.}

\item{.xvar}{독립변수. 독립변수가 여러 개일 때는 벡터 형태로 제공한다. (e.g. \code{c(age, height)}).}

\item{.ncomp}{잠재변수의 수.}

\item{.center}{평균조정 여부. TRUE이면 각 독립변수에 대해 평균조정을 수행한다. 현재 구현상 TRUE만 가능.}

\item{.scale}{분산조정 여부. TRUE이면 각 독립변수에 대해 분산조정을 수행한다.}
}
\value{
리스트.
}
\description{
주어진 데이터에 대하여 부분최소자승 회귀모형을 추정한다.
}
\examples{
data(biometric, package = "dmtr")
fit <- fit_plsr(biometric, weight, c(age, height), .ncomp = 1L)

}
\keyword{partial-least-squares}
