% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{fn_sum_of_squares_ols}
\alias{fn_sum_of_squares_ols}
\title{회귀 제곱합}
\usage{
fn_sum_of_squares_ols(.beta, .x, .y)
}
\arguments{
\item{.beta}{계수 벡터.}

\item{.x}{독립변수 행렬.}

\item{.y}{종속변수 벡터.}
}
\value{
회귀제곱합.
}
\description{
주어진 회귀계수로부터 관측 데이터에 대한 회귀제곱합을 계산한다.
}
