% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gainchart_group1}
\alias{gainchart_group1}
\title{집단별 관측빈도.}
\format{
10개의 행과 2개의 열로 이루어진 데이터 프레임:
\describe{
  \item{k}{집단 번호 (1, 2, ..., 10)}
  \item{n}{범주 1의 빈도}
}
}
\usage{
gainchart_group1
}
\description{
\code{\link{gainchart_yfreq}}의 1000개의 객체들에 대해 어떤 분류모형을 이용하여 범주 1(특정범주)에 대한
사후확률을 산출한 후 그 확률의 내림차순으로 전체 객체를 정렬하여 매 100개의 객체를 한 집단으로 구분하였을 때, 각 집단에 속하는 범주
1의 빈도.
}
\keyword{datasets}
