% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gainchart_yfreq}
\alias{gainchart_yfreq}
\title{실제범주별 관측빈도.}
\format{
3개의 행과 2개의 열로 이루어진 데이터 프레임:
\describe{
  \item{y}{범주 (1, 2, 3)}
  \item{n}{관측수}
}
}
\usage{
gainchart_yfreq
}
\description{
전체 객체 수가 1,000인 어떤 데이터의 실제 범주별 빈도.
}
\keyword{datasets}
