% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{gr_sum_of_squares_ols}
\alias{gr_sum_of_squares_ols}
\title{회귀 제곱합 미분값}
\usage{
gr_sum_of_squares_ols(.beta, .x, .y)
}
\arguments{
\item{.beta}{계수 벡터.}

\item{.x}{독립변수 행렬.}

\item{.y}{종속변수 벡터.}
}
\value{
회귀제곱합 편미분값.
}
\description{
회귀제곱합을 회귀계수에 대한 함수라 할 때, 각 회귀계수에 대하여 편미분한 함수의 값을 주어진 회귀계수에 대하여 계산한다.
}
\keyword{regression}
