% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{grad_binary_logistic_regression}
\alias{grad_binary_logistic_regression}
\title{이분 로지스틱 회귀분석 로그우도함수의 gradient.}
\usage{
grad_binary_logistic_regression(.betas, .x, .y)
}
\arguments{
\item{.betas}{계수 벡터.}

\item{.x}{독립변수 행렬.}

\item{.y}{종속변수 벡터.}
}
\value{
편미분값 벡터.
}
\description{
주어진 계수에서 우도함수의 편미분값을 계산한다.
}
