% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{healthbehavior}
\alias{healthbehavior}
\title{건강 문진}
\format{
3개의 행과 6개의 열로 이루어진 데이터 프레임:
\describe{
  \item{id}{환자 고유번호}
  \item{x1}{운동여부 (예: 1, 아니오: 0)}
  \item{x2}{음주여부 (예: 1, 아니오: 0)}
  \item{x3}{흡연여부 (예: 1, 아니오: 0)}
  \item{x4}{가족력여부 (예: 1, 아니오: 0)}
  \item{x5}{고혈압여부 (예: 1, 아니오: 0)}
}
}
\usage{
healthbehavior
}
\description{
3명에 대한 건강 관련 문진에 대한 답을 나타낸 자료
}
\keyword{datasets}
