% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris90}
\alias{iris90}
\title{붓꽃 데이터.}
\format{
90개의 행과 6개의 열로 이루어진 데이터 프레임:
\describe{
  \item{id}{관측치 고유번호}
  \item{x1}{Sepal.Length}
  \item{x2}{Sepal.Width}
  \item{x3}{Petal.Length}
  \item{x4}{Petal.Width}
  \item{class}{범주 (setosa, versicolor, virginica)}
}
}
\source{
\code{\link[datasets]{iris}}
}
\usage{
iris90
}
\description{
\code{\link[datasets]{iris}}의 150개 데이터 중 각 범주 별 30개의 관측치를 모아 총 90개의 관측치로 구성한
데이터.
}
\keyword{datasets}
