% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{loglik_binary_logistic_regression}
\alias{loglik_binary_logistic_regression}
\title{이분 로지스틱 회귀분석 로그우도함수.}
\usage{
loglik_binary_logistic_regression(.betas, .x, .y)
}
\arguments{
\item{.betas}{계수 벡터.}

\item{.x}{독립변수 행렬.}

\item{.y}{종속변수 벡터.}
}
\value{
로그우도함수값.
}
\description{
주어진 계수로부터 관측 데이터가 얻어질 우도함수값을 계산한다.
}
\keyword{logistic-regression}
