% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{mallows_c}
\alias{mallows_c}
\title{맬로우즈 Cp 통계량}
\usage{
mallows_c(.fit_reduced, .fit_full)
}
\arguments{
\item{.fit_reduced}{$p$항 회귀모형 추정 결과.}

\item{.fit_full}{완전모형 추정 결과.}
}
\value{
맬로우즈 Cp 통계량.
}
\description{
맬로우즈 Cp 통계량을 계산한다.
}
\examples{
data(biometric, package = "dmtr")
fit_full <- fit_linear_regression(biometric, weight, c(age, height))
fit_reduced <- fit_linear_regression(biometric, weight, height)
mallows_c(fit_reduced, fit_full)

}
