% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiplsreg}
\alias{multiplsreg}
\title{다종속변수 학습표본.}
\format{
7개의 행과 6개의 열로 이루어진 데이터 프레임:
\describe{
  \item{x1, x2, x3, x4}{4개의 독립변수}
  \item{y1, y2}{2개의 종속변수}
}
}
\usage{
multiplsreg
}
\description{
두 개의 종속변수를 지닌 학습표본.
}
\keyword{datasets}
