% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal-component-analysis.R
\name{nipals_pca}
\alias{nipals_pca}
\title{NIPALS 알고리즘 - 주성분분해}
\usage{
nipals_pca(X, .pc = NULL)
}
\arguments{
\item{X}{독립변수 행렬.}

\item{.pc}{추출할 주성분의 수.}
}
\value{
리스트. \code{T}: 스코어 행렬, \code{V}: 고유벡터 행렬.
}
\description{
NIPALS 알고리즘을 이용하여 주성분 스코어 행렬과 고유벡터행렬을 구한다.
}
\keyword{principal-component-analysis}
