% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-least-squares.R
\name{nipals_plsr}
\alias{nipals_plsr}
\title{NIPALS 알고리즘 - 부분최소자승 회귀분석}
\usage{
nipals_plsr(X, y, ncomp = NULL)
}
\arguments{
\item{X}{독립변수 행렬.}

\item{y}{종속변수 벡터.}

\item{ncomp}{추출할 잠재변수 개수.}
}
\value{
리스트. \code{T}: 스코어 행렬, \code{W}: 가중치 행렬, \code{P}: 로딩행렬, \code{b} 회귀계수.
}
\description{
NIPALS 알고리즘을 이용하여 부분최소자승 회귀모형을 구한다.
}
