% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-least-squares.R
\name{nipals_plsr_n}
\alias{nipals_plsr_n}
\title{NIPALS 알고리즘 - 부분최소자승 회귀분석}
\usage{
nipals_plsr_n(X, Y, ncomp = NULL)
}
\arguments{
\item{X}{독립변수 행렬.}

\item{Y}{종속변수 행렬.}

\item{ncomp}{추출할 잠재변수 개수.}
}
\value{
리스트.
}
\description{
NIPALS 알고리즘을 이용하여 부분최소자승 회귀모형을 구한다.
}
