% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-utils.R
\name{normalize_to_prob}
\alias{normalize_to_prob}
\title{비음행렬의 값을 확률로 변환.}
\usage{
normalize_to_prob(p, small_p = NULL)
}
\arguments{
\item{p}{실수형 행렬. 모든 값은 0 이상이어야 한다.}

\item{small_p}{값 변환 전 행렬 \code{p}의 모든 원소값에 더할 작은 양의 실수.}
}
\value{
행렬.
}
\description{
행렬의 각 행이 합계 1의 값을 갖도록 값을 변환.
}
\examples{
x <- matrix(runif(15), nrow = 5)
normalize_to_prob(x)
normalize_to_prob(x, small_p = 0)

}
\keyword{feature-utils}
