% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcrdata}
\alias{pcrdata}
\title{주성분 회귀분석 학습표본.}
\format{
6개의 행과 4개의 열로 이루어진 데이터 프레임:
\describe{
  \item{x1, x2, x3}{3개의 독립변수}
  \item{y}{종속변수}
}
}
\usage{
pcrdata
}
\description{
하나의 종속변수와 세 개의 독립변수를 가진 학습표본.
}
\keyword{datasets}
