% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcusage1}
\alias{pcusage1}
\title{PC 사용 데이터 1.}
\format{
10개의 행과 4개의 열로 이루어진 데이터 프레임:
\describe{
  \item{id}{사용자 고유번호}
  \item{x1}{나이}
  \item{x2}{PC 경험연수}
  \item{x3}{주당 사용시간}
}
}
\usage{
pcusage1
}
\description{
가정에서 PC를 사용하는 10명에 대한 나이, PC 경험연수, 주당 사용시간을 나타낸 데이터.
}
\keyword{datasets}
