% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcusage2}
\alias{pcusage2}
\title{PC 사용 데이터 2.}
\format{
10개의 행과 3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{id}{사용자 고유번호}
  \item{x1}{PC 사용경력}
  \item{x2}{주당 사용시간}
}
}
\usage{
pcusage2
}
\description{
10명의 사람(객체)에 대한 PC 사용경력과 주당 사용시간을 나타낸 데이터.
}
\keyword{datasets}
