% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within-group.R
\name{pooled_variance}
\alias{pooled_variance}
\title{합동 분산-공분산행렬. Pooled variance.}
\usage{
pooled_variance(.data, .group_var, .xvar)
}
\arguments{
\item{.data}{관측 데이터 프레임. A raw data frame.}

\item{.group_var}{범주변수. A variable to group by.}

\item{.xvar}{분산-공분산 행렬에 포함될 변수. One or more variables separated by commas within \code{c()}. Each
variable needs to be a numeric column of \code{.data}.}
}
\value{
하나의 분산-공분산 행렬. A covariance matrix.
}
\description{
분산-공분산행렬이 범주에 관계없이 동일하다고 가정하고, 합동 분산-공분산행렬을 범주별로 주어진 관측치를 이용하여 추정한다.
\code{pooled_variance} returns common variance.
}
\examples{
data(binaryclass2, package = "dmtr")
pooled_variance(binaryclass2, class, c(x1, x2))

}
\keyword{within-group}
