% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{posterior_binary_logistic_regression}
\alias{posterior_binary_logistic_regression}
\title{이분 로지스틱 회귀분석 사후확률 추정.}
\usage{
posterior_binary_logistic_regression(.betas, .new_data, ..., .reflevel)
}
\arguments{
\item{.betas}{이분 로지스틱 회귀분석 계수}

\item{.new_data}{새 관측 데이터 프레임.}

\item{...}{범주 분류에 사용될 변수.}

\item{.reflevel}{이분 범주에서 1(positive) 값을 지닐 범주.}
}
\value{
사후확률.
}
\description{
주어진 계수를 이용하여 새 데이터에 대해 사후확률을 추정한다.
}
\examples{
data(student, package = "dmtr")
fit <- fit_binary_logistic_regression(student, y, x1, x2, x3, .reflevel = "우수")
posterior_binary_logistic_regression(fit$betas, student, x1, x2, x3, .reflevel = "우수")

}
