% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant-functions.R
\name{predict_da}
\alias{predict_da}
\title{새 데이터에 대한 범주 예측.}
\usage{
predict_da(
  .f,
  .new_data,
  ...,
  .include_score = FALSE,
  .include_posterior = FALSE,
  .include_class = TRUE
)
}
\arguments{
\item{.f}{판별함수 리스트.}

\item{.new_data}{새 관측 데이터 프레임.}

\item{...}{범주 분류에 사용될 변수.}

\item{.include_score}{TRUE이면 범주별 판별함수값을 결과에 저장. default = FALSE.}

\item{.include_posterior}{TRUE이면 사후확률값을 결과에 저장. default = FALSE.}

\item{.include_class}{TRUE이면 추정범주값을 결과에 저장. default = TRUE.}
}
\value{
데이터 프레임: 범주 추정.
}
\description{
범주별 판별함수를 이용하여 새 데이터에 대한 범주를 예측한다.
}
\examples{
data(binaryclass2, package = "dmtr")
f <- ld_fun(binaryclass2, class, x1, x2)
predict_da(f, binaryclass2, x1, x2, .include_posterior = TRUE)

}
