% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{predict_linear_regression}
\alias{predict_linear_regression}
\title{다중선형회귀모형 미래반응치 예측.}
\usage{
predict_linear_regression(.fit, .new_data, .xvar, .interval = 0)
}
\arguments{
\item{.fit}{회귀모형 추정 결과.}

\item{.new_data}{새 관측 데이터 프레임.}

\item{.xvar}{예측에 사용될 변수.}

\item{.interval}{예측구간. 0인 경우 예측구간을 구하지 않으며, 0 과 1 사이일 경우 \code{(100 * .interval)}\%의 예측구간을 구한다.}
}
\value{
애측값 데이터프레임.
}
\description{
주어진 계수를 이용하여 새 데이터에 대해 종속변수값을 예측한다.
}
\examples{
data(biometric, package = "dmtr")
fit <- fit_linear_regression(biometric, weight, c(age, height))
predict_linear_regression(fit, biometric, c(age, height))
predict_linear_regression(fit, dplyr::tibble(age = 40, height = 170), c(age, height), 0.95)

}
