% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal-component-analysis.R
\name{predict_pca}
\alias{predict_pca}
\title{주성분 스코어 계산.}
\usage{
predict_pca(.fit, .new_data)
}
\arguments{
\item{.fit}{주성분 분해 결과.}

\item{.new_data}{새 관측 데이터 프레임.}
}
\value{
주성분 스코어 데이터프레임.
}
\description{
주어진 주성분 분해 결과를 이용하여 새 데이터에 대해 주성분 스코어를 계산한다.
}
\examples{
data(biometric, package = "dmtr")
fit <- fit_pca(biometric, .pc = 2L)
predict_pca(fit, biometric)

}
\keyword{principal-component-analysis}
