% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purchasedata_reference}
\alias{purchasedata_reference}
\title{시장바구니 구매 데이터.}
\format{
3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{customer}{고객 고유번호}
  \item{item}{상품 고유번호}
  \item{purchase}{구매여부 (1: 구매)}
}
}
\usage{
purchasedata_reference
}
\description{
5명의 고객 각각이 7개의 상품 중 구매한 상품을 나타내는 데이터.
}
\keyword{datasets}
