% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purchasedata_target}
\alias{purchasedata_target}
\title{시장바구니 목표고객 구매 데이터.}
\format{
3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{customer}{고객 고유번호}
  \item{item}{상품 고유번호}
  \item{purchase}{기존구매여부 (1: 구매)}
}
}
\usage{
purchasedata_target
}
\description{
\code{\link{purchasedata_reference}}의 시장바구니 데이터를 이용하여 상품 추천을 제공할 목표고객의 기존 구매 데이터.
}
\keyword{datasets}
