% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant-functions.R
\name{qd_fun}
\alias{qd_fun}
\title{범주별 판별함수 - 이차 판별 분석.}
\usage{
qd_fun(.data, .group_var, .xvar)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.group_var}{범주변수.}

\item{.xvar}{범주 분류에 사용될 변수.}
}
\value{
범주별 판별 함수.
}
\description{
범주별 판별함수를 범주별 평균벡터와 범주별 분산-공분산행렬을 이용하여 정의한다.
}
\examples{
data(binaryclass2, package = "dmtr")
f <- qd_fun(binaryclass2, class, c(x1, x2))

}
