% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rollsteel}
\alias{rollsteel}
\title{열연코일 인장강도 데이터.}
\format{
10개의 행과 3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{ct}{권취온도}
  \item{thickness}{두께 (2 혹은 6, 단위: mm) - 범주형(factor) 변수}
  \item{ts}{인장강도}
}
}
\usage{
rollsteel
}
\description{
열연코일의 인장강도(TS)에 권취온도(CT)가 어떤 영향을 미치는가를 조사하기 위한 학습표본.
}
\keyword{datasets}
