% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{select_variables_stepwise}
\alias{select_variables_stepwise}
\title{다중회귀모형 단계적 선택방법을 통한 회귀모형 입력변수 선택}
\usage{
select_variables_stepwise(
  .data,
  .yvar,
  .xvar,
  .alpha_in = 0.05,
  .alpha_out = 0.1,
  .maxit = 100L,
  .verbose = TRUE
)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.yvar}{종속변수.}

\item{.xvar}{완전모형에 속할 독립변수. 독립변수가 여러 개일 때는 벡터 형태로 제공한다. (e.g. \code{c(age, height)})}

\item{.alpha_in}{변수 선택 단계에서 적용할 유의수준. Default: 0.05}

\item{.alpha_out}{변수 제거 단계에서 적용할 유의수준. \code{.alpha_in} 보다 크거나 같아야 한다. Default: 0.10}

\item{.maxit}{최대 iteration 수. Default: 100.}

\item{.verbose}{각 단계의 선택/제거 변수를 화면에 출력할 지 여부. Default: \code{TRUE}}
}
\value{
최종 선택된 독립변수 이름 벡터.
}
\description{
단계적 선택방법을 통해 독립변수를 선택한다.
}
\examples{
data(biometric, package = "dmtr")
select_variables_stepwise(biometric, weight, c(age, height))

}
\keyword{regression}
