% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{student}
\alias{student}
\title{학생 평점 학습표본.}
\format{
15개의 행과 5개의 열로 이루어진 데이터 프레임:
\describe{
  \item{id}{학생 고유번호}
  \item{x1}{아침식사 여부 (예: 1, 아니오: 0)}
  \item{x2}{수면시간}
  \item{x3}{서클활동시간}
  \item{y}{평점범주 (우수 혹은 보통)}
}
}
\usage{
student
}
\description{
기숙사에 거주하는 대학생 15명을 대상으로 얻은 생활습관에 대한 설문조사 결과와 평점(이분형 범주 - 우수/보통).
}
\keyword{datasets}
