% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telconnection}
\alias{telconnection}
\title{성능변수에 따른 통신 만족도.}
\format{
12개의 행과 3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{N}{회선잡음}
  \item{L}{소리크기 손실}
  \item{y}{만족도 범주 - 서열형(ordinal) (나쁨: 1, 보통: 2, 좋음: 3)}
}
}
\source{
\url{https://doi.org/10.1002/j.1538-7305.1976.tb02939.x}
}
\usage{
telconnection
}
\description{
벨 연구소에서 한 통신잔치에 대하여 실시한 조사 결과. 주요 성능변수인 회선잡음(circuit noise: N)과 소리크기
손실(loudness loss: L)이 이용자의 주관적인 만족도에 미치는 영향을 분석하기 위한 것이다. 만족도는 원결과(Cavanaugh,
1976)를 가공하여 세 범주로 분류하였다.
}
\keyword{datasets}
