% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telsatisfaction}
\alias{telsatisfaction}
\title{통신서비스 만족도.}
\format{
3개의 행과 6개의 열로 이루어진 데이터 프레임:
\describe{
  \item{id}{고객 고유번호}
  \item{x1}{성별 - 이분형 (여 혹은 남)}
  \item{x2}{나이 - 연속형}
  \item{x3}{직업 - 명목형}
  \item{x4}{월통신요금 - 연속형}
  \item{x5}{통신서비스 만족도 - 서열형 (1: 매우만족, 2: 만족, 3: 보통, 4: 불만족, 5: 매우불만족)}
}
}
\usage{
telsatisfaction
}
\description{
3명의 고객별 통신서비스 만족도. 혼합형 객체 데이터의 예.
}
\keyword{datasets}
