% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telsatisfaction_range}
\alias{telsatisfaction_range}
\title{통신서비스 만족도 변수 범위.}
\format{
2개의 행과 4개의 열을 지닌 데이터 프레임:
\describe{
  \item{range}{"min": 최소, "max": 최대}
  \item{x2, x4, x5}{각 변수에 해당하는 최소/최대값}
}
}
\usage{
telsatisfaction_range
}
\description{
\code{\link{telsatisfaction}}의 연속형 및 서열형 변수들에 대한 값의 범위 (최소/최대).
}
\keyword{datasets}
