% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trafficviolation}
\alias{trafficviolation}
\title{운전경력에 따른 사고횟수.}
\format{
8개의 행과 3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{id}{운전자 고유번호}
  \item{x1}{운전경력}
  \item{x2}{위반횟수}
}
}
\usage{
trafficviolation
}
\description{
8명에 대한 운전경력과 교통위반 횟수를 나타낸 표본.
}
\keyword{datasets}
