% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactiondata}
\alias{transactiondata}
\title{트랜잭션.}
\format{
21개의 행과 2개의 열로 이루어진 데이터 프레임:
\describe{
  \item{transaction_id}{트랜잭션 고유번호}
  \item{item}{항목}
}
}
\usage{
transactiondata
}
\description{
총 5개의 트랜잭션에서 각 트랜잭션에 포함된 항목들을 나타낸 long-form 데이터 프레임.
}
\keyword{datasets}
