% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactionseq}
\alias{transactionseq}
\title{고객 시퀀스.}
\format{
16개의 행과 3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{customer_id}{고객 고유번호}
  \item{transaction_seq}{고객별 트랜잭션 순서}
  \item{item}{항목}
}
}
\usage{
transactionseq
}
\description{
5명의 고객들의 시간에 따른 구매 행태를 나타낸 long-form 데이터 프레임.
}
\keyword{datasets}
