% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{ttest_linear_regression}
\alias{ttest_linear_regression}
\title{다중선형회귀모형 회귀계수 검정.}
\usage{
ttest_linear_regression(.fit)
}
\arguments{
\item{.fit}{회귀모형 추정 결과.}
}
\value{
회귀계수 검정 데이터프레임.
}
\description{
추정된 회귀모형을 이용하여 각 회귀계수의 유의성을 검정한다.
}
\examples{
data(biometric, package = "dmtr")
fit <- fit_linear_regression(biometric, weight, c(age, height))
ttest_linear_regression(fit)

}
\keyword{regression}
