% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{vif_linear_regression}
\alias{vif_linear_regression}
\title{다중회귀모형 분산팽창계수}
\usage{
vif_linear_regression(.data, .xvar)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.xvar}{완전모형에 속할 독립변수. 독립변수가 여러 개일 때는 벡터 형태로 제공한다. (e.g. \code{c(age, height)})}
}
\value{
\code{.xvar}의 각 변수조합에 대한 분산팽창계수.
}
\description{
다중공선성의 정도를 알아보기 위해 분산팽창계수를 계산한다.
}
\examples{
data(biometric, package = "dmtr")
vif_linear_regression(biometric, c(age, height))

}
