% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordtfidf}
\alias{wordtfidf}
\title{문서별 단어 TF-IDF 가중치.}
\format{
30개의 행과 3개의 열로 이루어진 데이터 프레임:
\describe{
  \item{document}{문서 고유번호}
  \item{word}{단어 고유번호}
  \item{tfidf}{TF-IDF 가중치}
}
}
\usage{
wordtfidf
}
\description{
5개 중요 단어의 문서 6건에 대한 TF-IDF 가중치.
}
\keyword{datasets}
