% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordweight}
\alias{wordweight}
\title{사용자 단어 가중치.}
\format{
5개의 행과 2개의 열로 이루어진 데이터 프레임:
\describe{
  \item{word}{단어 고유번호}
  \item{weight}{목표사용자 가중치}
}
}
\usage{
wordweight
}
\description{
\code{\link{wordtfidf}}의 5개 중요 단어에 대하여 평가된 목표사용자의 가중치. 본 가중치와
\code{\link{wordtfidf}}의 각 문서별 가중치를 비교하여 목표사용자에게 문서를 추천하고자 하는 목적이다.
}
\keyword{datasets}
