% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{fit_ordinal_logistic_regression}
\alias{fit_ordinal_logistic_regression}
\title{서열형 로지스틱 회귀분석 계수 추정.}
\usage{
fit_ordinal_logistic_regression(
  .data,
  .group_var,
  .xvar,
  .type = "cumulative",
  .control = list(fnscale = -1, maxit = 1000)
)
}
\arguments{
\item{.data}{관측 데이터 프레임.}

\item{.group_var}{범주변수.}

\item{.xvar}{범주 분류에 사용될 변수.}

\item{.type}{로짓모형의 종류. "cumulative": 누적 로짓모형, "adjacent": 인근범주 로짓모형}

\item{.control}{\code{\link[stats]{optim}} 함수 수행 시 \code{control} 파라미터 리스트.}
}
\value{
리스트. 최우추정 계수 행렬 \code{betas}. 헤시안 행렬 \code{hessian}.
}
\description{
주어진 데이터에 대해 서열형 로지스틱 회귀모형 계수를 추정한다.
}
\examples{
data(telconnection, package = "dmtr")
fit_ordinal_logistic_regression(
  telconnection, y, c(N, L)
)
fit_ordinal_logistic_regression(
  telconnection, y, d(N, L), .type = "adjacent"
)

}
